+++ cyberdrop_dl/cli/download.py
@@ -70,7 +70,7 @@ async def _post_runtime(manager: Manager) -> None:
         file_hashes = await manager.hasher.run()
         await manager.deduper.run(file_hashes)
 
-    if manager.config.sorting.sort_downloads:
+    if manager.config.sort.enabled:
         await manager.sorter.run()
 
     _check_partials_and_empty_folders(manager.config)
@@ -110,7 +110,7 @@ def download(
 
     config = Config.create(appdata, cli.config_file).update(config)
 
-    if not cli.fullscreen_ui or cli.config_file or config.sorting.sort_downloads:
+    if not cli.fullscreen_ui or cli.config_file or config.sort.enabled:
         cli.download = True
 
     manager = Manager(cli, appdata, config)
@@ -120,7 +120,7 @@ def download(
 
 def _check_ffmpeg(config: Config) -> None:
     errors: list[Exception] = []
-    if config.sorting.needs_ffmpeg:
+    if config.sort.needs_ffmpeg:
         exc = RuntimeError("Sorting media files requires 'ffmpeg' to be installed")
         exc.add_note("Disable sorting or install ffmpeg")
         errors.append(exc)
@@ -154,6 +154,6 @@ def _delete_empty_files(config: Config) -> None:
     logger.info("Deleting empty files and folders...")
     cleanup.rm_empty_dirs(config.download_folder)
 
-    sorted_folder = config.sorting.sort_folder
-    if sorted_folder and config.sorting.sort_downloads:
+    sorted_folder = config.sort.output_folder
+    if sorted_folder and config.sort.enabled:
         cleanup.rm_empty_dirs(sorted_folder)