+++ cyberdrop_dl/config/__init__.py
@@ -30,7 +30,7 @@ from .settings import (
     MediaDurationLimits,
     RateLimiting,
     RuntimeOptions,
-    Sorting,
+    Sort,
     UIOptions,
 )
 
@@ -75,7 +75,7 @@ class Config(BaseModel):
     proxy: HttpURL | None = None
     rate_limits: RateLimiting = Field(default_factory=RateLimiting)
     runtime: RuntimeOptions = Field(default_factory=RuntimeOptions)
-    sorting: Sorting = Field(default_factory=Sorting)
+    sort: Sort = Field(default_factory=Sort)
     ssl_context: Literal["truststore", "certifi", "truststore+certifi"] | None = "truststore+certifi"
     ui_options: UIOptions = Field(default_factory=UIOptions)
     user_agent: NonEmptyStr = "Mozilla/5.0 (X11; Linux x86_64; rv:150.0) Gecko/20100101 Firefox/150.0"